requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
email-validator==1.1.3
PyJWT==2.1.0
Flask-Limiter==2.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.1